task-02-AI Agent with External Tool Access/requirements.txt
langgraph
langgraph-checkpoint-sqlite
langchain
langchain-google-genai
google-generativeai
requests
pandas
streamlit
python-dotenv